Set up environment for testing, new testing modules, refactor express's logic out to app.js, server is just listening
