+++ package.json
@@ -25,6 +25,7 @@
     "babel-preset-es2015": "^6.9.0",
     "babel-preset-react": "^6.11.1",
     "bootstrap": "^3.3.7",
+    "chai": "^3.5.0",
     "eslint": "^2.9.0",
     "eslint-config-airbnb": "^9.0.1",
     "eslint-plugin-import": "^1.10.3",
@@ -32,6 +33,8 @@
     "eslint-plugin-react": "^5.2.2",
     "font-awesome": "^4.6.3",
     "jquery": "^3.1.0",
+    "mocha": "^3.0.2",
+    "supertest": "^2.0.0",
     "webpack-dev-server": "^1.14.1"
   },
   "dependencies": {
