+++ server/app.js
@@ -1,0 +1,28 @@
+const express = require('express');
+require('dotenv').config();
+const app = express();
+const path = require('path');
+const bodyParser = require('body-parser');
+const db = require('./db/db.js');
+const router = require('./routes');
+
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: true }));
+app.use(express.static(path.join(__dirname, '../client/public/dist/')));
+
+app.all('/api/*', function(req, res, next) {
+  res.header("Access-Control-Allow-Origin", "*");
+  res.header("Access-Control-Allow-Headers", "Cache-Control, Pragma, Origin, Authorization, Content-Type, X-Requested-With");
+  res.header("Access-Control-Allow-Methods", "GET, PUT, POST, OPTIONS");
+  if (req.method.toLowerCase() !== "options") {
+   return next();
+  }
+  return res.sendStatus(204);
+})
+app.use('/api', router);
+
+app.get('*', (req, res)=>{
+  res.sendFile(path.join(__dirname + '/../client/public/dist/index.html'));
+});
+
+module.exports = app;
